Online Model/requirements.txt

# online/requirements.txt
pylsl
numpy
scipy
mne
onnxruntime
websockets
fastapi
uvicorn
paho-mqtt
pyyaml